Re-use parseParentPattern logic
