+++ test/integration/goldens/logging/com/google/cloud/logging/v2/ConfigClient.java
@@ -1134,7 +1134,7 @@ public class ConfigClient implements BackgroundResource {
    *
    * <pre>{@code
    * try (ConfigClient configClient = ConfigClient.create()) {
-   *   String parent = BillingAccountName.of("[BILLING_ACCOUNT]").toString();
+   *   String parent = ProjectName.of("[PROJECT]").toString();
    *   for (LogSink element : configClient.listSinks(parent).iterateAll()) {
    *     // doThingsWith(element);
    *   }
@@ -1161,7 +1161,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   ListSinksRequest request =
    *       ListSinksRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setPageToken("pageToken873572522")
    *           .setPageSize(883849137)
    *           .build();
@@ -1188,7 +1188,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   ListSinksRequest request =
    *       ListSinksRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setPageToken("pageToken873572522")
    *           .setPageSize(883849137)
    *           .build();
@@ -1214,7 +1214,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   ListSinksRequest request =
    *       ListSinksRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setPageToken("pageToken873572522")
    *           .setPageSize(883849137)
    *           .build();
@@ -1484,7 +1484,7 @@ public class ConfigClient implements BackgroundResource {
    *
    * <pre>{@code
    * try (ConfigClient configClient = ConfigClient.create()) {
-   *   String parent = BillingAccountName.of("[BILLING_ACCOUNT]").toString();
+   *   String parent = ProjectName.of("[PROJECT]").toString();
    *   LogSink sink = LogSink.newBuilder().build();
    *   LogSink response = configClient.createSink(parent, sink);
    * }
@@ -1517,7 +1517,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   CreateSinkRequest request =
    *       CreateSinkRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setSink(LogSink.newBuilder().build())
    *           .setUniqueWriterIdentity(true)
    *           .build();
@@ -1545,7 +1545,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   CreateSinkRequest request =
    *       CreateSinkRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setSink(LogSink.newBuilder().build())
    *           .setUniqueWriterIdentity(true)
    *           .build();
@@ -2015,7 +2015,7 @@ public class ConfigClient implements BackgroundResource {
    *
    * <pre>{@code
    * try (ConfigClient configClient = ConfigClient.create()) {
-   *   String parent = BillingAccountName.of("[BILLING_ACCOUNT]").toString();
+   *   String parent = ProjectName.of("[PROJECT]").toString();
    *   for (LogExclusion element : configClient.listExclusions(parent).iterateAll()) {
    *     // doThingsWith(element);
    *   }
@@ -2042,7 +2042,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   ListExclusionsRequest request =
    *       ListExclusionsRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setPageToken("pageToken873572522")
    *           .setPageSize(883849137)
    *           .build();
@@ -2069,7 +2069,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   ListExclusionsRequest request =
    *       ListExclusionsRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setPageToken("pageToken873572522")
    *           .setPageSize(883849137)
    *           .build();
@@ -2097,7 +2097,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   ListExclusionsRequest request =
    *       ListExclusionsRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setPageToken("pageToken873572522")
    *           .setPageSize(883849137)
    *           .build();
@@ -2358,7 +2358,7 @@ public class ConfigClient implements BackgroundResource {
    *
    * <pre>{@code
    * try (ConfigClient configClient = ConfigClient.create()) {
-   *   String parent = BillingAccountName.of("[BILLING_ACCOUNT]").toString();
+   *   String parent = ProjectName.of("[PROJECT]").toString();
    *   LogExclusion exclusion = LogExclusion.newBuilder().build();
    *   LogExclusion response = configClient.createExclusion(parent, exclusion);
    * }
@@ -2389,7 +2389,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   CreateExclusionRequest request =
    *       CreateExclusionRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setExclusion(LogExclusion.newBuilder().build())
    *           .build();
    *   LogExclusion response = configClient.createExclusion(request);
@@ -2414,7 +2414,7 @@ public class ConfigClient implements BackgroundResource {
    * try (ConfigClient configClient = ConfigClient.create()) {
    *   CreateExclusionRequest request =
    *       CreateExclusionRequest.newBuilder()
-   *           .setParent(BillingAccountName.of("[BILLING_ACCOUNT]").toString())
+   *           .setParent(ProjectName.of("[PROJECT]").toString())
    *           .setExclusion(LogExclusion.newBuilder().build())
    *           .build();
    *   ApiFuture<LogExclusion> future = configClient.createExclusionCallable().futureCall(request);
