+++ test/integration/goldens/logging/com/google/cloud/logging/v2/ConfigClientTest.java
@@ -427,7 +427,7 @@ public class ConfigClientTest {
 
     CreateBucketRequest request =
         CreateBucketRequest.newBuilder()
-            .setParent(FolderLocationName.of("[FOLDER]", "[LOCATION]").toString())
+            .setParent(LocationName.of("[PROJECT]", "[LOCATION]").toString())
             .setBucketId("bucketId-1603305307")
             .setBucket(LogBucket.newBuilder().build())
             .build();
@@ -456,7 +456,7 @@ public class ConfigClientTest {
     try {
       CreateBucketRequest request =
           CreateBucketRequest.newBuilder()
-              .setParent(FolderLocationName.of("[FOLDER]", "[LOCATION]").toString())
+              .setParent(LocationName.of("[PROJECT]", "[LOCATION]").toString())
               .setBucketId("bucketId-1603305307")
               .setBucket(LogBucket.newBuilder().build())
               .build();
